Fix code style issues: remove relative imports, add commas, fix constructors' position, rename packages
