+++ lib/src/widgets/controls/button_size_type.dart
@@ -1,5 +1,4 @@
-import 'package:flutter/material.dart';
-import '../../../layout/layout_grid.dart';
+import 'package:admiralui_flutter/layout/layout_grid.dart';
 
 enum ButtonSizeType {
   small,
